test: split native proof invalidation regressions
